+++ sample/sample1-simple-usage/app.ts
@@ -1,6 +1,7 @@
 import "reflect-metadata";
 import {createConnection, useContainer} from "typeorm";
 import {Container} from "typedi";
+
 import {PostRepository} from "./repository/PostRepository";
 import {Post} from "./entity/Post";
