+++ src/decorators/InjectManager.ts
@@ -2,9 +2,9 @@ import {ConnectionManager} from "typeorm";
 import {Container} from "typedi";
 
 /**
- * Allows to inject an Connection using typedi's Container.
+ * Allows to inject an EntityManager using typedi's Container.
  */
-export function OrmConnection(connectionName: string = "default"): Function {
+export function InjectManager(connectionName: string = "default"): Function {
     return function(object: Object|Function, propertyName: string, index?: number) {
         Container.registerHandler({ object, index, propertyName, value: () => {
             const connectionManager = Container.get(ConnectionManager);
@@ -13,7 +13,13 @@ export function OrmConnection(connectionName: string = "default"): Function {
                   `Make sure you have created such connection. Also make sure you have called useContainer(Container) ` +
                   `in your application before you established a connection and importing any entity.`);
 
-            return connectionManager.get(connectionName);
+            const connection = connectionManager.get(connectionName);
+            const entityManager = connection.manager;
+            if (!entityManager)
+                throw new Error(`Entity manager was not found on "${connectionName}" connection. ` +
+                  `Make sure you correctly setup connection and container usage.`);
+
+            return entityManager;
         }});
     };
 }
